Add func inserting data received from AWS Lambda to DB & Add func calling SELECT statement of DB

+++ app/src/main/java/io/github/yunato/myankicard/model/entity/QACard.kt
@@ -1,8 +1,8 @@
 package io.github.yunato.myankicard.model.entity
 
-data class QACard(val timestamp: Int,
+data class QACard(val timestamp: Long,
                   val question: String,
                   val answer: String,
-                  val next_date: Int,
-                  val consecutive: Int,
+                  val next_date: Long,
+                  val consecutive: Long,
                   var is_correct: Boolean)
